Move proof loading logic to model layer
