+++ src/repositories/local.ts
@@ -1,18 +1,10 @@
 import type { Readable } from 'svelte/store'
 
-import type { Ref, Trait } from '../models'
 import type { Prestore } from '../stores'
 import * as Deduction from '../stores/deduction'
 import * as Src from '../stores/source'
 
-type Serializer<T> = [
-  (value: T) => string, // serializer
-  (string: string) => T | undefined, // deserializer; should be inverse to the serializer
-]
-
-type Serializers<T> = {
-  [K in keyof T]: Serializer<T[K]>
-}
+import { Serializers, prestore } from './serializers'
 
 export type Source<T> = {
   [K in keyof T]: Readable<T[K]>
@@ -97,72 +89,6 @@ function build<T extends Record<string, unknown>>(
   }
 }
 
-// This type lies, as JSON.parse(JSON.stringify(m)) !== m for non-primitive types
-// Use with caution
-function json<T>(): Serializer<T> {
-  return [JSON.stringify, JSON.parse]
-}
-
-// Traits take a up a _lot_ of space in storage, so we use a compressed encoding
-type TraitRow =
-  | [1, number, number, boolean, string, Ref[]]
-  | [0, number, number, boolean, number[], number[]]
-
-const traits: Serializer<Trait[]> = [
-  (values) =>
-    JSON.stringify(
-      values.map((t) => {
-        if (t.asserted) {
-          return [1, t.space, t.property, t.value, t.description, t.refs]
-        } else {
-          return [
-            0,
-            t.space,
-            t.property,
-            t.value,
-            t.proof.properties,
-            t.proof.theorems,
-          ]
-        }
-      }),
-    ),
-  (str) => {
-    const traits: Trait[] = []
-
-    if (!str) {
-      return traits
-    }
-
-    JSON.parse(str).forEach((row: TraitRow) => {
-      if (row[0] === 1) {
-        const [_, space, property, value, description, refs] = row
-        traits.push({
-          asserted: true,
-          space,
-          property,
-          value,
-          description,
-          refs,
-        })
-      } else if (row[0] === 0) {
-        const [_, space, property, value, properties, theorems] = row
-        traits.push({
-          asserted: false,
-          space,
-          property,
-          value,
-          proof: {
-            properties,
-            theorems,
-          },
-        })
-      }
-    })
-
-    return traits
-  },
-]
-
 export const spec: Spec<Prestore> = {
   // TODO: it is possible that we're reading a value persisted by an earlier
   // code version's serializer. To prevent that case, the deserializers should
@@ -170,15 +96,7 @@ export const spec: Spec<Prestore> = {
   // Question: if one of the fields fails to pass validations, should we still
   // do a partial load (and fallback on the default for only those fields), or
   // should it invalidate all stored fields?
-  serializers: {
-    properties: json(),
-    spaces: json(),
-    theorems: json(),
-    traits: traits,
-    source: json(),
-    sync: json(),
-    deduction: json(),
-  },
+  serializers: prestore,
   initial: {
     properties: [],
     spaces: [],
@@ -190,6 +108,6 @@ export const spec: Spec<Prestore> = {
   },
 }
 
-export default function create() {
-  return build<Prestore>(spec)
+export default function create(specfication = spec) {
+  return build<Prestore>(specfication)
 }
